+++ backend/src/main/java/agilementor/member/controller/AuthController.java
@@ -1,6 +1,6 @@
 package agilementor.member.controller;
 
-import agilementor.member.service.AuthService;
+import agilementor.member.service.AuthClientService;
 import agilementor.member.service.MemberService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
@@ -12,17 +12,17 @@ import org.springframework.web.bind.annotation.RequestParam;
 @RequestMapping("/api/auth")
 public class AuthController {
 
-    private final AuthService authService;
+    private final AuthClientService authClientService;
     private final MemberService memberService;
 
-    public AuthController(AuthService authService, MemberService memberService) {
-        this.authService = authService;
+    public AuthController(AuthClientService authClientService, MemberService memberService) {
+        this.authClientService = authClientService;
         this.memberService = memberService;
     }
 
     @GetMapping("/login")
     public String redirectGoogleLogin() {
-        return "redirect:" + authService.getAuthUrl();
+        return "redirect:" + authClientService.getAuthUrl();
     }
 
     @GetMapping("/login/code/google")
@@ -31,7 +31,7 @@ public class AuthController {
         String idToken;
 
         try {
-            idToken = authService.requestIdToken(code);
+            idToken = authClientService.requestIdToken(code);
         } catch (Exception e) {
             // todo: 로그인 실패 페이지 추가
             return "redirect:/";
